+++ cli/report_chart.go
@@ -55,7 +55,7 @@ func chartReportCommand(t *core.Track, options *filterOptions) *cobra.Command {
 				out.Err("failed to generate report: %s", err)
 				return
 			}
-			filters = append(filters, core.FilterByTime(filterStart, filterEnd))
+			filters = core.NewFilter(filters.Functions, filterStart, filterEnd)
 
 			reporter, err := core.NewReporter(t, options.projects, filters, options.includeArchived, start, filterEnd)
 			if err != nil {